+++ src/lib/lisp/builtin.c
@@ -45,6 +45,8 @@ make_car_call (void)
 
 #ifdef __CC65__
 #pragma code-name ("CODE_INIT")
+#pragma inline-stdfuncs (off)
+#pragma allow-eager-inline (off)
 #endif
 
 // Make objects for built-in procedures.