+++ src/lib/lisp/error.c
@@ -1,5 +1,7 @@
 #ifdef __CC65__
 #include <ingle/cc65-charmap.h>
+#pragma inline-stdfuncs (off)
+#pragma allow-eager-inline (off)
 #endif
 
 #include <limits.h>