+++ src/lib/lisp/eval.c
@@ -759,6 +759,8 @@ funcall ()
 
 #ifdef __CC65__
 #pragma code-name ("CODE_INIT")
+#pragma inline-stdfuncs (off)
+#pragma allow-eager-inline (off)
 #endif
 
 extern lispptr needle;